python/plugins: mocked socket.getaddrinfo to make tests non-reliant on an active internet connection
